+++ packages/reown_core/lib/reown_core.dart
@@ -11,6 +11,9 @@ export 'relay_client/relay_client_models.dart';
 export 'store/store_models.dart';
 export 'utils/errors.dart';
 export 'utils/utils.dart';
+export 'utils/evm_utils.dart';
+export 'utils/polkadot_utils.dart';
+export 'utils/solana_utils.dart';
 export 'utils/constants.dart';
 export 'utils/method_constants.dart';
 export 'utils/log_level.dart';
