+++ packages/reown_core/test/solana_signature_test.dart
@@ -8,7 +8,7 @@ void main() {
     test('should extract the transaction id from a solana transaction', () {
       final transaction =
           'AeJw688VKMWEeOHsYhe03By/2rqJHTQeq6W4L1ZLdbT2l/Nim8ctL3erMyH9IWPsQP73uaarRmiVfanEJHx7uQ4BAAIDb3ObYkq6BFd46JrMFy1h0Q+dGmyRGtpelqTKkIg82isAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAMGRm/lIRcy/+ytunLDm+e8jOW7xfcSayxDmzpAAAAAtIy17v5fs39LuoitzpBhVrg8ZIQF/3ih1N9dQ+X3shEDAgAFAlgCAAABAgAADAIAAACghgEAAAAAAAIACQMjTgAAAAAAAA==';
-      final signature = ReownCoreUtils.extractSolanaSignature(transaction);
+      final signature = SolanaChainUtils.extractSolanaSignature(transaction);
       expect(
         signature,
         '5XanD5KnkqzH3RjyqHzPCSRrNXYW2ADH4bge4oMi9KnDBrkFvugagH3LytFZFmBhZEEcyxPsZqeyF4cgLpEXVFR7',
@@ -21,7 +21,7 @@ void main() {
       final transaction =
           '+dGmyRGtpelqTKkIg82isAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAMGRm/lIRcy/+ytunLDm+e8jOW7xfcSayxDmzpAAAAAtIy17v5fs39LuoitzpBhVrg8ZIQF/3ih1N9dQ+X3shEDAgAFAlgCAAABAgAADAIAAACghgEAAAAAAAIACQMjTgAAAAAAAA==';
       expect(
-        () async => ReownCoreUtils.extractSolanaSignature(transaction),
+        () async => SolanaChainUtils.extractSolanaSignature(transaction),
         throwsA(isA<ArgumentError>()),
       );
     });
